requirements.txt
pywin32
screen-ocr[winrt]
customtkinter
pydirectinput-rgx
numpy
pillow
pyautogui
